I. Cleanup Makefiles -    Eclipse is no longer a required part of the build process!
Now 'make' does it all:
  o It starts with 'make compile' which does what Eclipse used to do,
    starting with running JavaCC, then copying over the files from src
    into build, in addition to a bunch of running javac.
  o Then it does the same 'make bin' that it used to do to make the
    compiled rules compiler.
  o Then it does the old 'ant jar'.

II. Bonus:  Consistency of Java use.

It used to be that the rulescompiler got built with the Java on your $PATH:
'java', where the dispatchifier and the rest of the interpreter got built with
the java at your $JAVA_HOME, as did the jar.  Now it all uniformly uses
your $JAVA_HOME - it assumes that it is a standard JDK and not just a JRE
at the $JAVA_HOME.

III. Another bonus, but this one is a tradeoff:  Makefile dependencies.

Make now knows that building the jar requires the rulescompiler to be
built and that building the rulescompiler requires the compilation to
have occurred.  Therefore, you can't 'only' build the jar.  This
should prevent errors but it does take slightly longer, especially
since the combination of make and ant means that it doesn't always
know when to call ant, so it always calls it --- of course ant then
works out dependencies and won't re-compile all those java files.